+++ src/ui/setup/pages/SocialFlow/types/auto-imports.d.ts
@@ -237,6 +237,7 @@ declare global {
   const useNow: typeof import('@vueuse/core')['useNow']
   const useObjectUrl: typeof import('@vueuse/core')['useObjectUrl']
   const useOffsetPagination: typeof import('@vueuse/core')['useOffsetPagination']
+  const useOnboardingStore: typeof import('../../../../../stores/onboarding')['useOnboardingStore']
   const useOnline: typeof import('@vueuse/core')['useOnline']
   const useOptionsStore: typeof import('../../../../../stores/options.store')['useOptionsStore']
   const usePageLeave: typeof import('@vueuse/core')['usePageLeave']
@@ -584,6 +585,7 @@ declare module 'vue' {
     readonly useNow: UnwrapRef<typeof import('@vueuse/core')['useNow']>
     readonly useObjectUrl: UnwrapRef<typeof import('@vueuse/core')['useObjectUrl']>
     readonly useOffsetPagination: UnwrapRef<typeof import('@vueuse/core')['useOffsetPagination']>
+    readonly useOnboardingStore: UnwrapRef<typeof import('../../../../../stores/onboarding')['useOnboardingStore']>
     readonly useOnline: UnwrapRef<typeof import('@vueuse/core')['useOnline']>
     readonly useOptionsStore: UnwrapRef<typeof import('../../../../../stores/options.store')['useOptionsStore']>
     readonly usePageLeave: UnwrapRef<typeof import('@vueuse/core')['usePageLeave']>